backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]>=2.7.4,<3.0.0
pydantic-settings==2.1.0
PyJWT==2.8.0
bcrypt==4.1.1
PyPDF2==3.0.1
python-multipart==0.0.6
loguru==0.7.2
langchain==0.3.20
langchain-openai==0.2.0
redis>=5.0.0
neo4j>=5.15.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
networkx>=3.0
openai>=1.35.0
fastapi-mail==1.4.1